+++ lib/buildkite/pipelines/pipeline.rb
@@ -5,42 +5,16 @@ require 'pathname'
 
 module Buildkite
   module Pipelines
-    class Pipeline
-      attr_reader :steps
-      attr_reader :plugins
-      attr_reader :templates
+    class Pipeline < Builder
+      attr_reader :plugins, :groups
 
       def initialize(definition = nil, &block)
         @env = {}
-        @steps = []
         @plugins = {}
-        @templates = {}
+        @groups = []
         @processors = []
-        @notify = []
 
-        instance_eval(&definition) if definition
-        instance_eval(&block) if block_given?
-      end
-
-      [
-        Steps::Block,
-        Steps::Command,
-        Steps::Input,
-        Steps::Trigger,
-      ].each do |type|
-        define_method(type.to_sym) do |template = nil, **args, &block|
-          add(type, template, **args, &block)
-        end
-      end
-
-      def notify(*args)
-        if args.empty?
-          @notify
-        elsif args.first.is_a?(Hash)
-          @notify.push(args.first.transform_keys(&:to_s))
-        else
-          raise ArgumentError, 'value must be hash'
-        end
+        super
       end
 
       def env(*args)
@@ -53,24 +27,6 @@ module Buildkite
         end
       end
 
-      def skip(template = nil, **args, &block)
-        step = add(Steps::Skip, template, **args, &block)
-        # A skip step has a nil/noop command.
-        step.command(nil)
-        # Always set the skip attribute if it's in a falsey state.
-        step.skip(true) if !step.get(:skip) || step.skip.empty?
-        step
-      end
-
-      def wait(attributes = {}, &block)
-        step = add(Steps::Wait, &block)
-        step.wait(nil)
-        attributes.each do |key, value|
-          step.set(key, value)
-        end
-        step
-      end
-
       def plugin(name, uri, version)
         name = name.to_s
 
@@ -79,18 +35,6 @@ module Buildkite
         end
 
         @plugins[name] = [uri, version]
-      end
-
-      def template(name, &definition)
-        name = name.to_s
-
-        if templates.key?(name)
-          raise ArgumentError, "Template already defined: #{name}"
-        elsif !block_given?
-          raise ArgumentError, 'Template definition block must be given'
-        end
-
-        @templates[name.to_s] = definition
       end
 
       def processors(*processor_classes)
@@ -109,31 +53,30 @@ module Buildkite
         @processors
       end
 
+      def group(name = nil, &block)
+        group = Group.new(name)
+
+        templates.each do |name, definition|
+          group.template(name, &definition)
+        end
+
+        group.instance_eval(&block)
+
+        @groups << group
+      end
+
       def to_h
         pipeline = {}
         pipeline[:env] = env if env.any?
         pipeline[:notify] = notify if notify.any?
-        pipeline[:steps] = steps.map(&:to_h)
+        pipeline[:steps] = groups.any? ? groups.map(&:to_h) : []
+        pipeline[:steps] += steps.map(&:to_h)
 
         Helpers.sanitize(pipeline)
       end
 
       def to_yaml
         YAML.dump(to_h)
-      end
-
-      private
-
-      def add(step_class, template = nil, **args, &block)
-        steps.push(step_class.new(self, find_template(template), **args, &block)).last
-      end
-
-      def find_template(name)
-        return unless name
-
-        templates[name.to_s] || begin
-          raise ArgumentError, "Template not defined: #{name}"
-        end
       end
     end
   end
